Spliting up the ech0160 models into separate files.
