+++ opengever/disposition/ech0160/model/__init__.py
@@ -1,0 +1,10 @@
+NOT_SPECIFIED = u'keine Angabe'
+
+
+from opengever.disposition.ech0160.model.document import Document  # noqa
+from opengever.disposition.ech0160.model.dossier import Dossier  # noqa
+from opengever.disposition.ech0160.model.file import File  # noqa
+from opengever.disposition.ech0160.model.folder import Folder  # noqa
+from opengever.disposition.ech0160.model.position import Position  # noqa
+from opengever.disposition.ech0160.model.repository import Repository  # noqa
+from opengever.disposition.ech0160.model.contentrootfolder import ContentRootFolder  # noqa
