+++ opengever/disposition/ech0160/model/file.py
@@ -1,0 +1,27 @@
+from opengever.disposition.ech0160.bindings import arelda
+from opengever.disposition.ech0160.utils import file_checksum
+import binascii
+import os.path
+
+
+class File(object):
+    """eCH-0160 dateiSIP"""
+
+    def __init__(self, toc, document):
+        self.file = document.obj.archival_file or document.obj.file
+        self.id = u'_{}'.format(binascii.hexlify(self.file._p_oid))
+        document.file_ref = self.id
+        self.document = document
+        self.filename = self.file.filename
+        self.filepath = self.file._blob.committed()
+
+        base, extension = os.path.splitext(self.filename)
+        self.name = 'p{0:06d}{1}'.format(toc.next_file, extension)
+        toc.next_file += 1
+
+    def binding(self):
+        datei = arelda.dateiSIP(id=self.id)
+        datei.name = self.name
+        datei.originalName = self.filename
+        datei.pruefalgorithmus, datei.pruefsumme = file_checksum(self.filepath)
+        return datei
